+++ README.md
@@ -86,9 +86,9 @@
     - [x] 정적팩토리 메서드 추가
     - [x] setter 제거
 - [x] DTO 만들기
-- [ ] 서비스 로직을 도메인 로직으로 이동
+- [x] 서비스 로직을 도메인 로직으로 이동
+    - [x] Domain 단위 테스트 추가
 - [ ] 메서드 이름 의미있게 짓기 (setter 메서드 네이밍 변경)
 - [ ] 일급콜렉션 추가
 - [ ] 모든 원시값과 문자열은 포장
-- [ ] Domain 단위 테스트 추가
