+++ src/main/java/kitchenpos/menu/domain/Product.java
@@ -1,6 +1,7 @@
 package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
+import java.util.Objects;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -22,6 +23,7 @@ public class Product {
     }
     
     private Product(String name, BigDecimal price) {
+        checkPrice(price);
         this.name = name;
         this.price = price;
     }
@@ -40,6 +42,12 @@ public class Product {
 
     public BigDecimal getPrice() {
         return price;
+    }
+    
+    private void checkPrice(BigDecimal price) {
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("상품은 0원 이상이어야 합니다");
+        }
     }
 
 }
