+++ src/main/java/kitchenpos/order/application/TableService.java
@@ -80,4 +80,10 @@ public class TableService {
         return OrderTableResponse.from(orderTableRepository.save(savedOrderTable));
     }
     
+    @Transactional(readOnly = true)
+    public OrderTable findById(Long orderTableId) {
+        System.out.println("여기왜안들어오징..");
+        return orderTableRepository.findById(orderTableId)
+                .orElseThrow(() -> new IllegalArgumentException("해당하는 주문 테이블이 없습니다"));
+    }
 }
